relocate tests to their own directory
Preparing for transitioning to google-test framework
