+++ tests/main.cpp
@@ -1,4 +1,4 @@
-#include "test/test_falcon.hpp"
+#include "test_falcon.hpp"
 #include <iostream>
 
 int
